- Restructuring shop grid component. - Displaying already saved custom links.
