+++ packages/grid/index.js
@@ -37,7 +37,11 @@ export default connect(
         isBusinessAccount: profile.business,
         isLockedProfile: state.profileSidebar.isLockedProfile,
         customLinksDetails: currentProfile.customLinksDetails,
+        maxCustomLinks: currentProfile.maxCustomLinks,
         publicGridUrl: `https://shopgr.id/${profile.serviceUsername}`,
+        hasCustomLinksFlip: state.appSidebar.user.features
+          ? state.appSidebar.user.features.includes('shopgrid_links')
+          : false,
       };
     }
     return {};
