+++ packages/grid/reducer.test.js
@@ -25,8 +25,8 @@ describe('reducer', () => {
             buttonColor: null,
             buttonType: null,
             customLinks: [],
-            maxCustomLinks: 3,
           },
+          maxCustomLinks: 3,
           gridPosts: [],
         },
       },
@@ -51,8 +51,8 @@ describe('reducer', () => {
             buttonColor: null,
             buttonType: null,
             customLinks: [],
-            maxCustomLinks: 3,
           },
+          maxCustomLinks: 3,
           gridPosts: [post],
           total: 1,
         },
@@ -81,8 +81,8 @@ describe('reducer', () => {
             buttonColor: null,
             buttonType: null,
             customLinks: [],
-            maxCustomLinks: 3,
           },
+          maxCustomLinks: 3,
           gridPosts: [],
           total: 0,
         },
